% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormorbidity_6m.R
\name{comorbidity_6m}
\alias{comorbidity_6m}
\title{6 months prior comorbidity flags}
\usage{
comorbidity_6m(original_data = analytic_30dcmp, month = 6)
}
\arguments{
\item{month}{how many months to look back for comorbidity, default is 6 months}
}
\value{

}
\description{
\enumerate{
\item Self merge the data
\item Identify the 6-month prior comorbidities
\item Merge back to main analytic file
\item Cut the data off Jan-June 2007 since no enough data to look back 6 months
}
}
