% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complication_flags.R
\name{complication_flags}
\alias{complication_flags}
\title{Complication Flags based on Procedure and Diagnosis ICD}
\usage{
complication_flags(
  std_data_root = wd$std_data_root,
  fac_codes_folder = "fac_clm_code",
  original_data = analytic_readmit,
  cmp_pr,
  cmp_dx,
  all_n
)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{fac_codes_folder}{folder name for fac_code}

\item{original_data}{data name for the data you want to add the info to}

\item{cmp_pr}{procedure ICD code list used to define complication}

\item{cmp_dx}{diagnosis ICD code list used to define complication}

\item{all_n}{number of digits to match with ICD code}
}
\value{
add two variables; complication and complication w/o POA
}
\description{
complication flag and complication excluding POA
}
\details{
POA var only available after 2010-01-01.(no poa var recorded before)
\enumerate{
\item Create the ICD9/10 Dx/PR to complications mapping format
\item Create the complication flags at the DX/PR code level (poa and w/o poa)
note: both DX and pr code are used to define complication,
however pr should be within 30days, dx has no time frame.
}
}
