% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormorbidity_6m.R
\name{cormorbidity_6m}
\alias{cormorbidity_6m}
\title{6 months prior comorbidty flags}
\usage{
cormorbidity_6m(original_data = analytic_30dcmp)
}
\arguments{
\item{original_data}{}
}
\value{

}
\description{
\enumerate{
\item Self merge the data
\item Identify the 6-month prior comorbidities
\item Merge back to main analytic file
\item Cut the data off Jan-June 2007 since no enough data to look back 6 months
}
}
