% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reoperation.R
\name{emergent_merpar_claim_ids_all_year}
\alias{emergent_merpar_claim_ids_all_year}
\title{Emergent MedPAR claim ids}
\usage{
emergent_merpar_claim_ids_all_year(
  std_data_root = wd$std_data_root,
  fac_clm_folder = "fac_clm"
)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{fac_clm_folder}{facility claim file folder name}
}
\value{

}
\description{
This is used to refine readmission; readmission should only be defined as
emergent admission after the index procedure. This helps exclude any planned visit after
the index surgery. Only need to rerun if MedPAR original data has changed
or added new MedPAR year data
}
