% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_clm_code.R
\name{fac_clm_code}
\alias{fac_clm_code}
\title{Facility Claim Code Standardization}
\usage{
fac_clm_code(
  year,
  data_file_name,
  schema,
  src_root,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{year of medicare data}

\item{data_file_name}{original MedPAR dataset names}

\item{schema}{defined in csv file. defult is "fac_clm"}

\item{src_root}{locations of MedPAR on database}

\item{mapping_data}{select medicare original vars to mapped vars}
}
\value{

}
\description{
long format of diagnosis and procedure code, code version, date
and POA
}
