% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_dx.R
\name{fac_dx}
\alias{fac_dx}
\title{Diagnosis Code from Facility Claim}
\usage{
fac_dx(
  original_data = analytic_demo,
  std_data_root = wd$std_data_root,
  year = 2007,
  fac_codes_folder = "fac_clm_code",
  fac_clm_folder = "fac_clm"
)
}
\arguments{
\item{original_data}{data name for the data you want to add the info to}

\item{std_data_root}{data path to the standardized data folder}

\item{year}{select one year to process, e.g. 2007}

\item{fac_codes_folder}{folder name for fac_code}

\item{fac_clm_folder}{folder name for fac_clm data (.csv format is a must)}
}
\value{

}
\description{
add facility code info to the analytic dataset
}
\details{
\enumerate{
\item read std fac_clm and fac_clm_code data;
\item join fac_clm and prof_clm by memberID and service date info.
Prof claim dates should be within facility claim dates
\item then join fac_clm_code with analytic dataset by memberID and facility claim ID (attach dx code info)
\item create admission_type(urgent/not), 30d death
\item facility claim code are saved by year within fac_clm_code folder (due to data size)
}
}
