% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_dx_elix.R
\name{fac_dx_elix}
\alias{fac_dx_elix}
\title{Elixhauser flgs based on Dx ICD code using ahrq map}
\usage{
fac_dx_elix(original_data)
}
\arguments{
\item{original_data}{data to add on}
}
\value{

}
\description{
main function is using ICD Package comorbid(, map = icd9_map_ahrq)
# ICD package map reference: https://rdrr.io/cran/icd/man/icd9_map_ahrq.html
}
