% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computer_config.R
\name{find_maize_folder}
\alias{find_maize_folder}
\title{Title: find brian geroge folder location}
\usage{
find_maize_folder(Brian_George_folder_loc)
}
\arguments{
\item{Brian_George_folder_loc}{}
}
\value{

}
\description{
Title: find brian geroge folder location
}
