% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_if_save_data.R
\name{helper_check_if_save_data}
\alias{helper_check_if_save_data}
\title{Title check if to overwrite existing data}
\usage{
helper_check_if_save_data(
  save_or_not,
  wd = NULL,
  rdata_name = NULL,
  csv_name = NULL
)
}
\arguments{
\item{save_or_not}{"yes" or "no"; save or not save data on maize drive; if no, then this helper file
will not pop up questions.}

\item{wd}{Where the data you want to save at}

\item{rdata_name}{the dataset you want to save, e.g. membership}

\item{csv_name}{what dataset name you want to save as? has to be \code{.csv}. e.g. "membership.csv"}
}
\value{

}
\description{
Title check if to overwrite existing data
}
