% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_info.R
\name{hospital_info}
\alias{hospital_info}
\title{hospital info using AHA data}
\usage{
hospital_info(aha_data, original_data)
}
\arguments{
\item{aha_data}{: American Hospital Association data}

\item{original_data}{: data to add on}
}
\value{

}
\description{
hospital info using AHA data
}
