% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_map.R
\name{import_map}
\alias{import_map}
\title{import pre-defined mapping schema}
\usage{
import_map(
  schema,
  source = "source_schema",
  src_data = import_src,
  mapping_data = import_mapping
)
}
\arguments{
\item{schema}{schema of the output file defined in the mapping csv files. for example ""membership" is a schema}

\item{source}{defualt "source_schema" to generate std medicare data.
"source_file": get medicare data files name; This option is used for standarzation process.
"source_schema": to get user defined schema name; only use to get all vars under the same
medicare dataset. This was used to create sas code.}

\item{mapping_data}{csv dataset that includes medicare original variable names to std data var names}

\item{import_src}{csv dataset that includes medicare original data to mapped std data}
}
\value{

}
\description{
Output are: medicare dataset name, standarzied dataset name, select varaible from medicare
and map to std variable names
}
\examples{

}
