% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fac_dx_code.R
\name{load_fac_code_data}
\alias{load_fac_code_data}
\title{load standardized fac clm code data across years}
\usage{
load_fac_code_data(std_data_root, fac_codes_folder)
}
\arguments{
\item{fac_codes_folder}{}
}
\value{

}
\description{
load standardized fac clm code data across years
}
