% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{read in membership data (demon/MBSF)}
\usage{
membership(
  year,
  data_file_name,
  schema,
  src_root,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{year of medicare data}

\item{data_file_name}{names of the original medicare dataset}

\item{schema}{schema of the output file defined in the mapping csv files. for example ""membership" is a schema}

\item{src_root}{the original medicare data folder location}

\item{mapping_data}{csv dataset that includes medicare original variable names to std data var names}
}
\value{

}
\description{
just combine all years bene to one dataset and keep unique record by "member_id"
and "member_year"
}
\examples{

}
