% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_data_dir_path.R
\name{mk_data_dir_path}
\alias{mk_data_dir_path}
\title{Make working directory locations}
\usage{
mk_data_dir_path(data_root = "sample", medicare_file, george_file)
}
\arguments{
\item{data_root}{: "sample" or "full_data"}

\item{medicare_file}{: original medicare data file folder name. e.g. demon data folder
name is "Denom_MBSF"}

\item{george_file:}{where the BG folder located at on maize. You may use the \code{computer_config}
to help you detect.}
}
\value{
list items of data directory locactions.
}
\description{
working directory location may change based on the computers and file locations;
This was based on the mapped Maize location on Xilin Chen's Mac and Windows
}
