% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_surgeon_proce_assist.R
\name{multi_surgeon_proc_assit_flags}
\alias{multi_surgeon_proc_assit_flags}
\title{Multiple suregons, procedures and assistant surgeons flags}
\usage{
multi_surgeon_proc_assit_flags(original_data)
}
\arguments{
\item{original_data}{data to add flags on}
}
\value{

}
\description{
\enumerate{
\item One admission has been operated by multiple physicians;
\item One admission has multiple operations;
\item Assistant physicians (using cpt mod code 80|82)
}
}
