% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_selection.R
\name{procedure_selection}
\alias{procedure_selection}
\title{Select Procedures from Professional Claim data}
\usage{
procedure_selection(
  std_data_root = wd$std_data_root,
  prof_codes_folder = "prof_clm",
  cpt_map = define_proc_by_cpt,
  year = 2007,
  test_sas_processed_data_loc = NA,
  n_worker = 2
)
}
\arguments{
\item{std_data_root}{professional claim data location}

\item{prof_codes_folder}{professional claim folder name}

\item{cpt_map}{should include 3 variable: cpt_cd, e_proc_grp and e_proc_grp_lbl}

\item{year}{select one year to process, e.g. 2007}

\item{test_sas_processed_data_loc}{this is for testing, only assign value when you want to test,
eg. compare the processing results with SAS code Zhaohui Fan created.
Otherwise leave it NA. When testing, assign the location of sas pre processed prof_clm data.
eg. "/Volumes/George_Surgeon_Projects/medicare_data/sample_npct_std/prof_clm.sas7bdat"}

\item{n_worker}{number of workers to use in future map}
}
\value{

}
\description{
select procedure related claims using CPT
}
\details{
\enumerate{
\item select procedures base on CPT
\item delete missing NPI procedures(can't link to surgeons)
\item delete duplicated claims
\item aggregate cpt_mode (if two cases have same info but diff cpt_mode, aggregate as one case include two cpt_mode code)
}
}
