% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof_clm_code.R
\name{prof_clm_code}
\alias{prof_clm_code}
\title{Title: professional claim diagnosis code}
\usage{
prof_clm_code(
  year,
  schema,
  data_file_name_clm,
  src_root_clm,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{year of medicare}

\item{schema}{defined in mapping csv, eg "prof_clm1"}

\item{data_file_name_clm}{original carrier claim dataset names}

\item{src_root_clm}{location to carrier claim dataset}
}
\value{

}
\description{
Title: professional claim diagnosis code
}
\details{
process using carrier claim; produce long format dx icd and icd version
}
