% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professionals.R
\name{professionals}
\alias{professionals}
\title{Title: read in physicains data}
\usage{
professionals(
  year,
  data_file_name,
  schema,
  src_root,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{: year of medicare data}

\item{data_file_name}{: original medicare dataset names}

\item{schema}{: defined in csv file. defult is "professional"}

\item{src_root}{: locations of carrier line on database}
}
\value{
combined physician NPI across years
}
\description{
Title: read in physicains data
}
\details{
: used in step3_professional
combine all years physician to one dataset and delete who have missing NPI
}
