% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{read_file}
\alias{read_file}
\title{Read analytic file}
\usage{
read_file(
  std_analytic_root = wd$std_analytic_root,
  folder_name = NULL,
  year,
  ...
)
}
\arguments{
\item{...}{from fread}
}
\value{

}
\description{
Read analytic file
}
