% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmission.R
\name{readmission}
\alias{readmission}
\title{Readmission Flags}
\usage{
readmission(
  original_data,
  std_data_root = wd$std_data_root,
  fac_clm_folder = "fac_clm",
  emergent_admission_medpar_claim_ids,
  year = 2007,
  max_year = 2018
)
}
\arguments{
\item{original_data}{data to add on}

\item{std_data_root}{data path to the standardized data folder}

\item{fac_clm_folder}{facility claim folder name}

\item{emergent_admission_medpar_claim_ids}{a list of all emergent MedPAR admission claim IDs}

\item{year}{select one year to process, e.g. 2007}

\item{max_year}{the most recent medicare year we have; for example, in 2020, we have Medpar processed up to 2018}
}
\value{

}
\description{
using facility claim data, left join by member id
}
\details{
\enumerate{
\item Find the 30-day readmission,  maybe multiple readmission
\item Roll up to a single row per member_id & prof_calim_id
\item Merge the 30-day readmission flag back to the main analytic file
\item readmission to emergent inpatient admission
}
}
