% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reoperation.R
\name{reoperation}
\alias{reoperation}
\title{30 days Re-operation Flags}
\usage{
reoperation(
  original_data,
  reop_map,
  std_data_root = wd$std_data_root,
  year,
  fac_codes_folder = "fac_clm_code",
  max_year = 2018
)
}
\arguments{
\item{original_data}{data to add on}

\item{reop_map}{predefined re-operation and ICD map}

\item{std_data_root}{data path to the standardized data folder}

\item{year}{select one year to process, e.g. 2007}

\item{fac_codes_folder}{names of the fac_clm_code data folder; or .sas files for only testing}

\item{max_year}{the most recent medicare year we have; for example, in 2020, we have Medpar processed up to 2018}
}
\value{

}
\description{
use procedure code ICD code to define reop
}
\details{
\enumerate{
\item Looking through the procedure codes across multiple admissions;
\item Create the ICD9/10 procedure codes to re-operation mapping using pre-defined operation types;
\item Get the re-operation flags if defined re operation happened within <=30 days
}
}
