% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_or_full_data_check.R
\name{sample_or_full_data_check}
\alias{sample_or_full_data_check}
\title{Title check if user want to use smaple or full data}
\usage{
sample_or_full_data_check(use_sample_or_not)
}
\arguments{
\item{use_sample_or_not}{: two option "sample" or "full_data"}
}
\value{

}
\description{
Title check if user want to use smaple or full data
}
\details{
to avid accidental full data processing, there will be a popup question to
confirm if choing full data
}
