% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{save_file}
\alias{save_file}
\title{Save files in create analytic file step uisng fwrite}
\usage{
save_file(
  data,
  std_analytic_root = wd$std_analytic_root,
  save_folder = NULL,
  year
)
}
\arguments{
\item{data}{file name to save}

\item{save_folder}{name of the folder to save under}

\item{year}{year will be used as the csv data name}

\item{wd}{directory path to the analytic file folder}
}
\value{

}
\description{
Save files in create analytic file step uisng fwrite
}
