% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_selection.R
\name{select_procedure}
\alias{select_procedure}
\title{Select Procedures from Professional Claim data}
\usage{
select_procedure(
  std_data_root = wd$std_data_root,
  professional_clm_data_name = "prof_clm.csv",
  cpt_map = define_proc_by_cpt
)
}
\arguments{
\item{std_data_root}{profesisonal claim data location}

\item{professional_clm_data_name}{profesisonal claim data name; has to be csv format}

\item{cpt_map}{should include 3 variable: cpt_cd, e_proc_grp and e_proc_grp_lbl}
}
\value{

}
\description{
Select Procedures from Professional Claim data
}
\details{
select procedure realted claims using CPT
use in creating analytic file
}
