% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_info.R
\name{ses_info}
\alias{ses_info}
\title{social economic status based on zip code}
\usage{
ses_info(ses = ses, ses_zscore = NULL, original_dataset = analytic_elix)
}
\arguments{
\item{ses}{input ses dataset name}

\item{ses_zscore}{input ses dataset name; leave as NULL if you don't want to add Zscore}

\item{original_dataset}{data to add on}
}
\description{
social economic status based on zip code
}
