% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severe_cm_flg.R
\name{severe_cmp_flags}
\alias{severe_cmp_flags}
\title{Severe Complication flags}
\usage{
severe_cmp_flags(original_data, perc = 0.75)
}
\arguments{
\item{perc}{percentage cut of point for length of stay considered severe}
}
\value{

}
\description{
Severe Complication flags
}
