% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bene_info.R
\name{bene_info}
\alias{bene_info}
\title{Patient demographics}
\usage{
bene_info(std_data_root, original_data, year)
}
\arguments{
\item{std_data_root}{data path to the membership info folder}

\item{original_data}{data name for the data you want to add the patient info on}

\item{year}{select one year to process, e.g. 2007}
}
\description{
Add patient demographics to the analytic file
}
\details{
\enumerate{
\item Add dob, dod, gender and race
\item Limit the patient population to age 65-99
}
}
