% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bene_info.R
\name{bene_info}
\alias{bene_info}
\title{Patient demographics}
\usage{
bene_info(std_data_root, member_data_name, original_data)
}
\arguments{
\item{std_data_root}{data path to the membership info folder}

\item{member_data_name}{data name for the membership dataset}

\item{original_data}{data name for the data you want to add the patient info to}
}
\description{
Add patient demographics to the analytic file
}
\details{
Add dob, dod, gender and race; Limit the patient population to age 65-99
}
