% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_pr_dx_code_map.R
\name{cmp_pr_dx_code_map}
\alias{cmp_pr_dx_code_map}
\title{Procedure and Diagnosis code that defines complication}
\usage{
cmp_pr_dx_code_map(complication_map = cmp_define)
}
\arguments{
\item{complication_map}{: pre-defined csv input data names located at input data folder
"icd910_dx_pr_2cmp_fmt.sas7bdat"}
}
\value{
cmp_pr = ICD9/10 for procedures;
cmp_dx = ICD9/10 for diagnosis;
all_n = number of digits to match with
}
\description{
create procedure and diagnosis icd list using for complication definition
}
\details{
use for define complications (complication_flags)
}
