% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_clm.R
\name{fac_clm}
\alias{fac_clm}
\title{facility claim standardization}
\usage{
fac_clm(year, data_file_name, schema, src_root, mapping_data = import_mapping)
}
\arguments{
\item{year}{year of medicare data}

\item{data_file_name}{original MedPAR dataset names}

\item{schema}{defined in csv file. defult is "fac_clm"}

\item{src_root}{locations of MedPAR on database}

\item{mapping_data}{select medicare original vars to mapped vars}
}
\value{
combined facility claims over years (no dup)
}
\description{
combines all years MedPAR claims (not including dx and pr code),
and removed the duplication claims;
}
