% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_clm.R
\name{fac_clm}
\alias{fac_clm}
\title{Title: facility claim standardized data
details: this scrip combines all years MedPPAR claims (not include dx and pr code),
and removed the duplication claims; for example, service date, facility NPI}
\usage{
fac_clm(year, data_file_name, schema, src_root)
}
\arguments{
\item{year}{: year of medicare data}

\item{data_file_name}{: original MedPAR dataset names}

\item{schema}{: defined in csv file. defult is "fac_clm"}

\item{src_root}{: locations of MedPAR on database}
}
\value{
combined facility claims over years; Delete duplications
}
\description{
Title: facility claim standardized data
details: this scrip combines all years MedPPAR claims (not include dx and pr code),
and removed the duplication claims; for example, service date, facility NPI
}
