% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_dx.R
\name{fac_dx}
\alias{fac_dx}
\title{Diagnosis Code from Facility Claim}
\usage{
fac_dx(
  std_data_root = wd$std_data_root,
  fac_codes_folder = "fac_clm_code",
  fac_clm_name = "fac_clm.csv",
  original_data = analytic_demo
)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{fac_codes_folder}{folder name for fac_code}

\item{fac_clm_name}{data name for fac_clm data}

\item{original_data}{data name for the data you want to add the info to}
}
\value{

}
\description{
add facility code info to the analytic dataset
}
\details{
read std fac_clm and fac_clm_code data; join fac_clm and prof_clm by memberID
and service date info; create admission_type(urgent/not), 30d death; attach dx code info
}
