% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilities.R
\name{facilities}
\alias{facilities}
\title{Combine Facility NPI across years}
\usage{
facilities(
  year,
  data_file_name,
  schema,
  src_root,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{: year of MEDPAR data}

\item{data_file_name}{: MEDPAR dataset names}

\item{schema}{: user defined in the mapping csv files. default is "facilities"}

\item{src_root}{: MedPAR dataset locations}

\item{mapping_data:}{select medicare original vars to mapped vars}
}
\value{

}
\description{
Combine Facility NPI across years
}
\details{
combine years facility NPI and delete duplications
}
