% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computer_config.R
\name{find_maize_folder}
\alias{find_maize_folder}
\title{Title: find brian geroge folder location
This script detects the computer enviromnent automatically}
\usage{
find_maize_folder(your_BG_folder_loc = NA)
}
\arguments{
\item{your_BG_folder_loc}{put your own BG location if the function can't find yours. e.g."/Volumes/George_Surgeon_Projects/"}
}
\value{
Brian_George folder location on maize, if it can't automaticly detect the location, you
can also manually specify the location
}
\description{
Title: find brian geroge folder location
This script detects the computer enviromnent automatically
}
\examples{



}
