% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fac_dx_code.R
\name{load_fac_code_data}
\alias{load_fac_code_data}
\title{load standardized fac clm dx code across years}
\usage{
load_fac_code_data(std_data_root, fac_codes_folder, year)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{fac_codes_folder}{folder name for fac_code}

\item{year}{select one year to process, e.g. 2007}
}
\value{

}
\description{
load standardized fac clm dx code across years
}
\details{
filter to only diagnosis code
}
