% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_data_dir_path.R
\name{mk_data_dir_path}
\alias{mk_data_dir_path}
\title{Make working directory locations}
\usage{
mk_data_dir_path(
  data_root = "sample",
  medicare_file = NULL,
  george_file,
  medicare_data_file = "/original_medicare_selected_vars/data",
  medicare_std_data_file = "/standardized_medicare_data_using_R",
  input_data_file = "/standardized_medicare_data_using_R/input/"
)
}
\arguments{
\item{data_root}{: "sample" or "full_data"}

\item{medicare_file}{: source (raw) medicare data file folder name. e.g. Denom_MBSF folder;
This is used to create source medicare data location.}

\item{george_file}{Brian_George folder loc on Maize}

\item{medicare_data_file}{original medicare location}

\item{medicare_std_data_file}{where the std medicare data will be saved to location}

\item{input_data_file}{location of outsource datasets, e.g. npppes npi files.}
}
\value{
list items of data directory locactions.
}
\description{
working directory location may change based on the computers and file locations;
This was based on the mapped Maize location on Xilin Chen's Mac and Windows
}
\examples{
\dontrun{
working_dir_config(data_root = "sample",
                   medicare_file = "Denom_MBSF",
                   george_file = "/Volumes/George_Surgeon_Projects")
}
}
