% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_selection.R
\name{procedure_selection}
\alias{procedure_selection}
\title{Select Procedures from Professional Claim data}
\usage{
procedure_selection(
  std_data_root = wd$std_data_root,
  professional_clm_data_name = "prof_clm.csv",
  cpt_map = define_proc_by_cpt
)
}
\arguments{
\item{std_data_root}{professional claim data location}

\item{professional_clm_data_name}{professional claim data name; has to be csv format}

\item{cpt_map}{should include 3 variable: cpt_cd, e_proc_grp and e_proc_grp_lbl}
}
\value{

}
\description{
select procedure related claims using CPT
}
\details{
\enumerate{
\item select procedure based on CPT
\item delete missing NPI procedures(can't link to surgeons)
\item delete duplicated claims
\item aggregate cpt_mode (if two cases have same info but diff cpt_mode, aggregate as one case include two cpt_mode code)
}
}
