% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof_clm.R
\name{prof_clm}
\alias{prof_clm}
\title{Title professional claim data}
\usage{
prof_clm(
  year,
  schema,
  data_file_name_clm,
  data_file_name_ln,
  src_root_clm,
  src_root_ln
)
}
\arguments{
\item{schema}{defined in csv mapping files, e.g. "prof_clm1"}

\item{data_file_name_clm}{carrier claim file names}

\item{data_file_name_ln}{carrier line file names}

\item{src_root_clm}{carrier claim file loc}

\item{src_root_ln}{carrier claim line loc}
}
\value{

}
\description{
Title professional claim data
}
\details{
use carrier claim and carrier line file. Physician and procedure
info is located at carrier line file.
}
