% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof_clm.R
\name{prof_clm}
\alias{prof_clm}
\title{Professional claim data}
\usage{
prof_clm(
  year,
  schema,
  data_file_name_clm,
  data_file_name_ln,
  src_root_clm,
  src_root_ln,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{year of medicare}

\item{schema}{defined in csv mapping files, e.g. "prof_clm1"}

\item{data_file_name_clm}{carrier claim file names}

\item{data_file_name_ln}{carrier line file names}

\item{src_root_clm}{carrier claim file loc}

\item{src_root_ln}{carrier claim line loc}

\item{mapping_data}{select medicare original vars to mapped vars}
}
\value{

}
\description{
use carrier claim and carrier line files. Physician and procedure
info is located at carrier line file.
}
\details{
variables included are:     "member_id", "claim_id", "clm_from_dt", "clm_thru_dt",
"clm_disp_cd", "icd_dx_prncpal", "icd_dx_prncpal_vrsn"
}
