% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof_clm_code.R
\name{prof_clm_code}
\alias{prof_clm_code}
\title{professional claim diagnosis code}
\usage{
prof_clm_code(year, schema, prof_clm, mapping_data = import_mapping)
}
\arguments{
\item{year}{year of medicare}

\item{schema}{defined in mapping csv, eg "prof_clm1"}

\item{prof_clm}{carrier claim data}
}
\value{

}
\description{
process using carrier claim; produce long format dx icd and icd version
}
