% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professionals.R
\name{professionals}
\alias{professionals}
\title{Read in physicians npi data}
\usage{
professionals(
  year,
  data_file_name,
  schema,
  src_root,
  mapping_data = import_mapping
)
}
\arguments{
\item{year}{year of medicare data}

\item{data_file_name}{original medicare dataset names}

\item{schema}{defined in csv file. default is "professional"}

\item{src_root}{locations of carrier line on database}

\item{mapping_data:}{select medicare original vars to mapped vars}
}
\value{
combined physician NPI across years
}
\description{
used in step3_professional
combine all years physician to one dataset and delete who have missing NPI
}
\details{
1.read all NPIs info across years from carrier line file
2. rename NPI variable
3. Combine all NPIs and delete NA values
}
