% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmission.R
\name{readmission}
\alias{readmission}
\title{Readmission Flags}
\usage{
readmission(std_data_root, fac_clm_name, original_data)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{fac_clm_name}{facility claim dataset name, ".csv"; you can use ".sas7bdat" data too.
But it is much slower due to read_sas is slow for big data.}

\item{original_data}{data to add on}
}
\value{

}
\description{
using facility claim data, left join by member id
}
\details{
\enumerate{
\item Find the 30-day readmission,  maybe multiple readmission
\item Roll up to a single row per member_id & prof_calim_id
\item Merge the 30-day readmission flag back to the main analytic file
}
}
