% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reoperation.R
\name{reoperation}
\alias{reoperation}
\title{30 days Re-operation Flags}
\usage{
reoperation(
  std_data_root = wd$std_data_root,
  reop_map,
  fac_codes_folder = "fac_clm_code",
  original_data
)
}
\arguments{
\item{std_data_root}{data path to the standardized data folder}

\item{reop_map}{predefined re-operation and ICD map}

\item{fac_codes_folder}{names of the fac_clm_code data folder}

\item{original_data}{data to add on}
}
\value{

}
\description{
use procedure code ICD code to define reop
}
\details{
\enumerate{
\item Looking through the procedure codes across multiple admissions;
\item Create the ICD9/10 procedure codes to re-operation mapping using pre-defined operation types;
\item Get the re-operation flags one row per professional claim for each patient;
\item Merge the re-operation flags back to main analytic file;
}
}
